minor bug fix
The last update I disabled the part of make_user.sh that fixed file paths pertaining to the user's home directory. This was done due to the length of time it was taking.

I noticed a bug this created in Nemo post-install and added a line which will fix the bug, without taking forever.