update manual partitioner window Manual partitioner now uses drop-down menus. This:  - Avoids confusion as to what needs to be entered  - Prevents possible typos when typing in partition names  - Allows for prevention of installation attempts to certain file systems or partitions which are too small
This functionality is not fully working just yet. Acceptance function needs to be modified to accept new widgets and possibly perform fewer checks.
Further, several bugs still need to be worked out, such as the loss of settings when drives are rescanned.
Finally, this needs review from others to ensure usage is smooth and understandable.