Add KDE support to boot-time Immersion
This will hide the KDE panel at boot time. This is not complete, we need to disable desktop icons too. However, it's a good deal better than what we had.

This also allows the $PATH binary to be simpler, not hooking into the Python interpreter at all. We had to move all the Python code it was running into engine.py

Signed-off-by: Thomas Castleman <contact@draugeros.org>