Add light error handling for mounting partitions
Before, we kinda just sent it: no checking. No retries. No errors if mounting failed. Nothing.

This commit has __mount__() checking if the designated device is mounted at the designated location, and if it isn't, retrying. If it fails a certain number of times, __mount__() throws an error and the entire installation fails with an error message.