Attempted major performance optimization
Discovered this while trying to get Gemini to help me optimize this in other ways.

We were just spawning all processes and then cleaning them all up. For some of these, which are IO-bound (making swap, making the user, setting the time), that's fine. But, some of these can be heavily multithreaded (updates, installing packages). So, instead, only spawn the same number of processes as we have of CPU cores.