Better error handling for updates
For some reason, the python library `apt` keeps failing to do it's job. So, all the other processing is still needed, we can just call `apt-get update`/`apt-get install`/`apt-get upgrade` instead, if the library throws an error.

Is it less elegant? Yes.
Is it less Pythonic? Yes.
Does it work? Also yes.